+++ apps/studio/src/components/Searchbar/index.ts
@@ -1,0 +1,1 @@
+export { Searchbar } from "./Searchbar"
